# perf: improve dot distance computing