--- a/rust/lance-index/benches/4bitpq_dist_table.rs
+++ b/rust/lance-index/benches/4bitpq_dist_table.rs
@@ -9,6 +9,7 @@ use arrow_array::types::Float32Type;
 use arrow_array::{FixedSizeListArray, UInt8Array};
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use lance_arrow::FixedSizeListArrayExt;
+use lance_index::vector::pq::distance::{build_distance_table_dot, build_distance_table_l2};
 use lance_index::vector::pq::ProductQuantizer;
 use lance_linalg::distance::DistanceType;
 use lance_testing::datagen::generate_random_array_with_seed;
@@ -21,7 +22,54 @@ const PQ: usize = 96;
 const DIM: usize = 1536;
 const TOTAL: usize = 16 * 1000;
 
-fn dist_table(c: &mut Criterion) {
+fn construct_dist_table(c: &mut Criterion) {
+    let codebook = generate_random_array_with_seed::<Float32Type>(256 * DIM, [88; 32]);
+    let query = generate_random_array_with_seed::<Float32Type>(DIM, [32; 32]);
+
+    c.bench_function(
+        format!(
+            "construct_dist_table: {},PQ={}x{},DIM={}",
+            DistanceType::L2,
+            PQ,
+            4,
+            DIM
+        )
+        .as_str(),
+        |b| {
+            b.iter(|| {
+                black_box(build_distance_table_l2(
+                    codebook.values(),
+                    4,
+                    PQ,
+                    query.values(),
+                ));
+            })
+        },
+    );
+
+    c.bench_function(
+        format!(
+            "construct_dist_table: {},PQ={}x{},DIM={}",
+            DistanceType::Dot,
+            PQ,
+            4,
+            DIM
+        )
+        .as_str(),
+        |b| {
+            b.iter(|| {
+                black_box(build_distance_table_dot(
+                    codebook.values(),
+                    4,
+                    PQ,
+                    query.values(),
+                ));
+            })
+        },
+    );
+}
+
+fn compute_distances(c: &mut Criterion) {
     let codebook = generate_random_array_with_seed::<Float32Type>(256 * DIM, [88; 32]);
     let query = generate_random_array_with_seed::<Float32Type>(DIM, [32; 32]);
 
@@ -38,7 +86,7 @@ fn dist_table(c: &mut Criterion) {
         );
 
         c.bench_function(
-            format!("{},{},PQ={},DIM={}", TOTAL, dt, PQ, DIM).as_str(),
+            format!("{},{},PQ={}x{},DIM={}", TOTAL, dt, PQ, 4, DIM).as_str(),
             |b| {
                 b.iter(|| {
                     black_box(pq.compute_distances(&query, &code).unwrap());
@@ -53,12 +101,12 @@ criterion_group!(
     name=benches;
     config = Criterion::default().significance_level(0.1).sample_size(10)
         .with_profiler(PProfProfiler::new(100, Output::Flamegraph(None)));
-    targets = dist_table);
+    targets = construct_dist_table, compute_distances);
 
 #[cfg(not(target_os = "linux"))]
 criterion_group!(
     name=benches;
     config = Criterion::default().significance_level(0.1).sample_size(10);
-    targets = dist_table);
+    targets = construct_dist_table, compute_distances);
 
 criterion_main!(benches);