for 8bit:
                        time:   [582.61 µs 585.08 µs 587.42 µs]
                        change: [-9.7495% -6.2223% -2.3288%] (p = 0.01 < 0.10)
                        Performance has improved.

for 4bit:
                        time:   [549.20 µs 551.13 µs 553.55 µs]
                        change: [-38.117% -37.734% -37.350%] (p = 0.00 < 0.10)
                        Performance has improved.